requirements.txt
# image2text.py requierements

pdf2image==1.16.3
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
matplotlib==3.8.2
